// src/api/products/api.ts
export interface ProductResponse {
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'reference'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'name'?: string;
    /**
     * 
     * @type {number}
     * @memberof ProductResponse
     */
    'price'?: number;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'observation'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'expiresAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'createdAt'?: string;
    /**
     * 
     * @type {number}
     * @memberof ProductResponse
     */
    'taxes'?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductResponse
     */
    'descount'?: number;
}